[Feature] Create MantraPaceControl component
